DS-3563: Use SumEmptyHandling (#9)
Use SumEmptyHandling() in place of Sum() when expecting the sum of an empty vector to yield a value of 0:
* comparingTwoPointsInTime
* revenueCalculation
* GrowthAccounting
* processMergers
* createFilters

Further, fixed a typo in GrowthAccounting where the "== 0" had been included within the summation, rather than comparing the result of the summation to zero. This was introduced in the previous replacement of sum() with Sum().